Move the user creation form
I am moving all user management handling to a single page to hopefully
improve the UX. I noticed that there was a lot of unnecessary steps that
had to be made just to add a user to the system. As I hadn't added the
rest of the management functionality it seemed wise to change things
pre-emptively.