Implement functionality for deleting SSH keys
A temporary file is created and the lines we want to keep are dumped
into this file. Once this process is finished and the key we want to
delete has not been copied to the temp file the file is closed. At which
point the file is moved to overwrite the original file.

Also included is a method for caching the file data which also handles
separating each key.