Remove ruby version requirement from Gemfile
The presence of this version specification prevents the server
executing commands with any ruby version other that the given (here
2.5.0) - instead throwing an exception.

This functionality runs contrary to our goals as we require internal
execution of tools held by `flight runway` which has its own embedded
ruby executable of version 2.6.0 (at time of writing).

Note: The root level `.ruby-version` file is fine as that only
specifies the version of ruby to actually run the server with.