Refactor network management page
This page functions mostly the same. However initially the user is shown
a card for each available network. These cards display relevant
information on their respective network based on the output of the
global `network show` script.

On the top right hand side of each card is an edit button. This allows
the user to view and edit the network variables associated with the
network on the card. This is essentially what this page showed before.