Adjust width and recenter body
To fit with other pages I have adjusted the width of the body content.
Full width pages are arguably not the best for reading content as nobody
likes a wall of text. There is probably further improvement to make on
the listing of assets to combat this but I haven't had much time to
think about this and for now I am happy with how that works for the
initial release.