Authenticate via PAM
Users are not stored within the DB in the traditional manner. Now when a
user logs in via a PAM user the username is saved to the DB and nothing
else.

This requires a PAM user to exist with a password set which is used to
login on the root page.