Refactor user view
The index for the user view has been split into multiple files to make
things easier to manage going forward. Each section is now a separate
file that the index calls to render.

On top of this the page itself has been refactored to fit more in line
with the style of the root page. This means the user creation form and
existing users table are now housed within neighbouring cards.