Fix button deconfigure button sizing
This resolves the button rendering incorrectly on certain screen sizes.

Note: There are some edge cases where this is still a problem but I find
it unlikely that this will cause any headaches for any particular
device. This is only a problem when resizing it to an obscure screen
size.

Sidenote: That note is going to come back and bite me sometime down the
line so I apologise to my future self, slightly.